Refactor monolithic NEAT script into a distributable Python package.
This commit refactors the initial `neat_xor.py` script into a structured, installable Python library named `neat_ml`.

Key changes include:
- Decomposing the core logic into modules: `neat`, `genome`, `problem`, `persistence`, `visualization`, and `history`.
- Creating a `pyproject.toml` to define the package, dependencies, and a `neat-train` console script.
- Implementing an abstract `Problem` class to allow developers to define their own problems for the NEAT algorithm to solve.
- Providing a clean set of developer examples in the `examples/` directory.
- Adding a comprehensive `README.md` with installation and usage instructions.
- Including a `.gitignore` file to exclude generated artifacts.
